Set up startRemoveExpense
Whenever we refreshed the webapp after removing an expense and refreshing the dashboard page,
the removed expense would still show up because the expense wasn't removed from Firebase.
I created startRemoveExpense in order to take care of that problem. Before deploying to Heroku, I made a test case for startRemoveExpense.
Changes have been made to the Edit page and it's tests to use startRemoveExpense instead of removeExpense.